+++ rmit-combined/recipes/configure.rb
@@ -3,3 +3,4 @@ cwb_benchmark_suite 'rmit-benchmark-suite'
 # Implementations provided under `files/default/benchmark_name.rb`
 cwb_benchmark 'sysbench-cpu'
 cwb_benchmark 'sysbench-memory'
+cwb_benchmark 'sysbench-threads'