+++ rmit-combined/CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.1.1 (2017-03-26)
 
-* Add sysbench: cpu, memory
+* Add sysbench: cpu, memory, threads
 
 # 0.1.0 (2017-03-25)